+++ dart/planning/RRT.h
@@ -90,11 +90,11 @@ public:
 public:
 
 	//// Constructor with a single root 
-    RRT(simulation::World* world, dynamics::Skeleton* robot, const std::vector<int> &dofs, const Eigen::VectorXd &root,
+		RRT(simulation::World* world, dynamics::Skeleton* robot, const std::vector<size_t> &dofs, const Eigen::VectorXd &root,
 			double stepSize = 0.02);
 
 	/// Constructor with multiple roots (so, multiple trees)
-    RRT(simulation::World* world, dynamics::Skeleton* robot, const std::vector<int> &dofs,
+		RRT(simulation::World* world, dynamics::Skeleton* robot, const std::vector<size_t> &dofs,
 			const std::vector<Eigen::VectorXd> &roots, double stepSize = 0.02);
 
 	/// Destructor
@@ -144,8 +144,8 @@ public:
 protected:
 
 	simulation::World* world;                 ///< The world that the robot is in
-    dynamics::Skeleton* robot;        ///< The ID of the robot for which a plan is generated
-	std::vector<int> dofs;                    ///< The dofs of the robot the planner can manipulate
+	dynamics::Skeleton* robot;        ///< The ID of the robot for which a plan is generated
+	std::vector<size_t> dofs;                    ///< The dofs of the robot the planner can manipulate
 
 	/// The underlying flann data structure for fast nearest neighbor searches 
 	flann::Index<flann::L2<double> >* index;
